fix: remove circular logic call between constants & util
There appears to have been a circular import call between constants and util. This was resolved by breaking the utils file into a directory of files. In this way, the functions that do not have any imports from constants.ts are resolved before the functions that require an import from constants.ts