feat: prepare repo to be published on npm (#25)
* chore: add base functions from contracts-v2

* fix: condense and enforce typing

* improve: add updated eslint & ignore  bindings

* fix: add new line for EOL

* fix: add additional ignore dirs

* fix: modify package & tsconfig files

* fix: remove circular logic call between constants & util

There appears to have been a circular import call between constants and util. This was resolved by breaking the utils file into a directory of files. In this way, the functions that do not have any imports from constants.ts are resolved before the functions that require an import from constants.ts

* feat: add index file to export types

* fix: revert changes that do not contribute

* fix: revert target version no

* nit: update comment content